web/requirements.txt
django==4.0.4
gunicorn==20.1.0
requests==2.27.1
Pillow==9.1.1
